Refactored serializers to not use Base64

+++ src/main/java/net/dashflight/data/caching/CacheStore.kt
@@ -5,7 +5,7 @@ interface CacheStore {
 
     fun set(key: String, value: ByteArray): Boolean
 
-    fun setWithExpiry(key: String, value: ByteArray, seconds: Int): Boolean
+    fun setWithExpiry(key: String, value: ByteArray, seconds: Long): Boolean
 
     fun get(key: String): ByteArray?
